Fix type declarations by moving them back to index
The generated type declarations did not work after tsc. This fixes it
